feat: 添加资源配置

+++ repository/src/main/java/com/SReMake/repository/system/ResourcesRepository.java
@@ -3,5 +3,5 @@ package com.SReMake.repository.system;
 import com.SReMake.model.system.Resources;
 import org.babyfish.jimmer.spring.repo.JavaRepository;
 
-public interface ResourcesRepository extends JavaRepository<Resources, Integer> {
+public interface ResourcesRepository extends JavaRepository<Resources, Long> {
 }
